backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
fastf1>=3.4.0
numpy>=1.24.0
pandas>=2.0.0
websockets>=12.0
python-dotenv>=1.0.0
